refactor: extract startup logic
