+++ roapi-http/src/lib.rs
@@ -3,6 +3,7 @@
 pub mod api;
 pub mod config;
 pub mod error;
+pub mod startup;
 
 #[cfg(test)]
 pub mod test_util;
